chore: move commonware node
